Alldata debug logic abstracted out
