+++ app.py
@@ -15,6 +15,7 @@ login_manager = LoginManager()
 login_manager.init_app(app)
 
 # These are here to avoid circular imports
+from brain.admin.all_data import AllData
 from brain.user_management.loginerator import Loginerator
 from brain.user_management.login_results import LoginResults
 from brain.user_management.register_city import RegisterCity
@@ -25,14 +26,7 @@ from models import Users, RepExercisesHistory, RepExercisesTaxonomy
 @app.route('/')
 @login_required
 def index():
-    user_results = list(Users.query.all())
-    taxonomy_results = list(RepExercisesTaxonomy.query.all())
-    taxonomy_results = [_prepare_taxonomy_entry(x) for x in taxonomy_results]
-    history_results = list(RepExercisesHistory.query.all())
-    history_results = [_prepare_history_entry(x) for x in history_results]
-    entries = [user_results, taxonomy_results, history_results]
-
-    return render_template('index.html', entries=entries)
+    return render_template('index.html', entries=AllData.get_all_data())
 
 
 @app.route('/login', methods=['GET', 'POST'])
@@ -75,35 +69,6 @@ def register():
 @login_manager.user_loader
 def user_loader(user_email):
     return Users.query.get(user_email)
-
-
-def _prepare_history_entry(entry):
-    """
-    Stringifies date of RepExercisesHistory entry
-    """
-    entry.date = str(entry.date)
-    return entry
-
-
-def _prepare_taxonomy_entry(entry):
-    """
-    Turns boolean True or False to YES or NO for RepExercisesTaxonomy entry
-    """
-    def debooleanize(value):
-        return 'YES' if value else 'NO'
-
-    entry.is_back = debooleanize(entry.is_back)
-    entry.is_chest = debooleanize(entry.is_chest)
-    entry.is_shoulders = debooleanize(entry.is_shoulders)
-    entry.is_biceps = debooleanize(entry.is_biceps)
-    entry.is_triceps = debooleanize(entry.is_triceps)
-    entry.is_legs = debooleanize(entry.is_legs)
-    entry.is_core = debooleanize(entry.is_core)
-    entry.is_balance = debooleanize(entry.is_balance)
-    entry.is_cardio = debooleanize(entry.is_cardio)
-    entry.is_weight_per_hand = debooleanize(entry.is_weight_per_hand)
-
-    return entry
 
 
 if __name__ == '__main__':
