+++ brain/user_management/loginerator.py
@@ -10,7 +10,7 @@ class Loginerator(object):
     """
     Manages the logic of logging in.
 
-    I N T E R F A C E   G U A R E N T E E D
+    I N T E R F A C E   G U A R A N T E E D
     ---------------------------------------
     login(email, password):
         -- Logs in a user if user exists and password is correct
